refactor: Move default providers implementation to according package
